Prevent cursor icon from getting stuck as the wrong shape
Once its gotten close to the edge of the window, my cursor will usually
get stuck as one of the resize arrows rather than change back to a
normal cursor.  It looks like Substance gets confused about what the
last cursor was, and starts "restoring" one of the resize cursors,
instead of the pointer, when the cursor moves back away from the edge of
the window. Setting the cursor for the frame's LayeredPane fixes the
problem.