Make idle notifier work for splashing/missing
When you splash or miss too many times HP dissapears and thus client
thinks that you have -1 HP and then idle notifier ignores that. Reason
why -1 check was there is because it was added in modification when we
were preventing the NPCs to throw notifications. But after some testing
it looks like that just checking for combat level is enough.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>