Team Capes: Added Team cape renaming
Parses input to reassign team cape names. example screenshot: https://i.imgur.com/L4O36ll.png if some of the input is invalid, just skips and continues parsing for other assignments. If the name is too long, it will keep it saved but output the max possible numbers of characters for the name to not clip with the cape count. Also slightly changed the client API's PanelComponent to make some positioning constants public, and added a getter for the width, to help make precise formatting.