Add a couple sound constants for later plugin use
I want to make runelite quack at me. So I found the quack sound ID by listening to the sounds one by one.

When exploring, some sounds are corrupted and do not end properly, so after a few you have to restart the whole client.
The sounds added here are safe to use and do end.

Also I discovered playing sound effects with `client.playSoundEffect` only is observable if the player has the volume for the primary sounds above 0.
These do not play in the environmental / area sound channel.