Sort Fishing spots by distance from camera
The default sort of fishing spots ends up drawing further things on top
of closer things, as it has no relation to the camera.

To solve this, convert the camera point to a `LocalPoint` and then
sort by comparing local points from the NPC to the camera's point.
However, the sort is inverted by `-1 *` so that further things are drawn
first, and closer things are drawn last--this way closer things are
always on top.

Fixes #1737