Use proper object id when possible.
Also fixes some WorldPoint positions for the ones I've changed. Many of them seem to have been added in the wrong direction. For example, if I'm standing at X = 1000 and the object is to my right, they'd put 999 instead of 1001.

I also changed some of the `NULL_*` names in ObjectID related to spirit trees. One of the clues needed ID 1293, which was named `NULL_1293` but referenced the spirit tree in Tree Gnome Village.