Maintain Agility plugin state on plugin state change
Clearing the state caused the plugin to lose information about already spawned objects which are necessary to display the overlays correctly.
Since the overlays are not used when the plugin is disabled anyway, we simply maintain the state as well.

Note that this still loses information about any spawned objects while the plugin was disabled, however fixing that would require a far more drastic rewrite of this plugin.