Raids Plugin Extension for Olm Crystal Bombs AoE Indicator + Timer
Added a tracker for crystal bombs AoE w/ timer indicators that functions similarly to the feature utilised by Konduit. In addition to this, the indicator polygons change colours depending on the proximity of the player to the center of each bomb, with green signifying safe, and other colours representing various increments of damage on a yellow-red spectrum.