Add quest filtering plugin
This plugin allows for users to filter their quest list by type (f2p, members, & miniquests), as well as by progress status (completed, in progress, and not started). Any combination of the 6 filtering options is valid.

The filtering is done during the initialization of the quest list widget (script 1350.rs2asm AKA QuestListInit.rs2asm) and during the initialization of each quest list section (script 1351.rs2asm AKA QuestListItemInit.rs2asm). A side effect of this is that the quest list must be reinitialized in order for configuration changes to be applied. This can be done simply by switching to the achievement diary list and back to the quest tab list. Whenever a quest list filter is changed through the configuration menu, a text panel will appear and give reinitialization instructions to the user. Once the list is reinitialized, the text panel will disappear and the selected filters will have been applied.