Make filtering instant and remove dependency on custom 1350 script
These changes will make all filtering of the quest list instantaneous. Any time a filtering config is changed, a call is made to script 1350 to reinitialize the quest list. This call is made from within the QuestTabPlugin class.

All logic from the custom 1350 script - QuestListInit.rs2asm - has been moved to the custom 1351 script - QuestListItemInit.rs2asm. Since 1350 is updated every time a quest is added and 1351 is not, removing dependency on the custom 1350 script will greatly reduce the amount of maintenance required to keep the plugin working.