Add mouse hover changes to various ui elements
This commit adds the following with the main aim of improving the overall ui experience:
InfoPanel: changes the panels to act more like buttons, they highlight on hover and the cursor changes to a pointer (since they are hyperlinks).
NewsFeedPanel: changes the cursor to a pointer on hover to indicate a hyperlink
HighscorePanel: changes the buttons below the search bar to highlight on hover to help indicate a clickable element.
FarmingPanel: same as above expect for the patch buttons
GE Panel: same as above but for the offer and search button
ConfigPanel: Changes the cog color on hover to help indicate clickable element.