Add world map region overlay dev tool
Using the existing 'World Map Location' dev tool, the world map can now
be given an overlay to show all the region boundaries and region ids.
This should make it easier to find out the region id of a specific
location.

One caveat is that this does not work with 'stitched' locations,
such as Motherlode Mine.