Added Olm Crystal Phase Bomb Visual Support
I inserted some additional code into the RaidsPlugin and RaidsConfig elements, as well as added two additional classes to allow for visual tracking of the Great Olm's crystal AoE bomb attack. This visually manifests in a similar manner to that of Konduit's plugin, with the additional feature of having a colour code associated with the closeness of the player to the center of the bomb (e.g. green being safe, red being directly on top of it, and variations in between being differing levels of severity/damage). This was done to offer additional support in raids for players.