config: add ability to stop sync
- Add sync keyword to config interface to allow for disabling syncing of
config properties. There is an icon for when properties are not supposed to be synced.
- This requires calls to setConfiguration to pass whether the config
item syncs.
- THE LOCAL PROPERTY FILE IS NOW LOADED BEFORE LOADING THE SYNCED
VERSION. This was done because the file gets deleted when calling load().
The local unsynced keys would get deleted. Now the properties instead get
updated by the synced properties, and unsynced keys remain untouched.