Addressed Change Requests
Extra tab of indentation fixed.
Bombs removed on events: despawn and gamestate changed to LOGGED_IN. (did this using a helper method in RaidsPlugin, not sure if this is the intended change.)
Utilized getTickCount() to track the bomb's up time, still having to use Instant.now() on each tick + each client update to generate the smooth countdown though.
(e.g. 1.2, 1.1, 1.0, etc. vs 1.8, 1.6, 1.2)
Altered the onGameTick() event to account for these changes.
Removed enum Normal_Color and replaced it with private static final Strings for each of the hexadecimal colours.
Bails out in the render override if config bombDisplay is off.
Comments alligned with code for next line comments.