Added boolean onSelectedEvent to Material Tabs
This new functionality allows a tab to not be selected depending on the
value of onTabSelected.

For example, for the hiscores, I want to block tab switching when the
plugin is already loading someone's hiscores, and in this scenario, I
have to return false to make sure the tabs don't switch.