Fix NMZ points overlay appearing in the KBD instance lair.
This was due to the old method checking to see if the only map tile loaded is 9033, which is both the KBD lair and NMZ (early OSRS reuses a lot). This doesn't cause issues with the normal KBD lair, as there are (mostly) empty but valid regions around the KBD lair.