Expand player indicators plugin configurations
A drop-down menu is introduced for each category of players, allowing
the choice of highlighting players, having them appear on the minimap,
or both.

Ref: https://user-images.githubusercontent.com/2199511/41439109-078c404a-7019-11e8-965b-79cbfc15b03d.png

Note: This will mean that the "Draw names on minimap" option must be
checked along with setting the desired player categories to "Both" in
order for player names of that category to appear on the minimap.

Closes runelite/runelite#3592