Adds a chat message that shows the layout when starting the raid
https://i.imgur.com/yuj4jEO.png 
Little QOL for Chambers of Xeric so you can keep track of the rotation and know what's next. I usually write it in the chat but this is a a bit better. I tested it on 20+ raids like in the picture, writing the rotations beforehand. Included a toggle which is default to on. I'm not a big fan of the text color either but it's to keep it consistent with the rest of the raid's text, and it will contrast with the regular black game text once you actually start doing stuff.