Added message on raid entrance & type indicators for Unknown rooms
Added the option so you can see the layout appear on entrance/scout (before the raid start):
![image](https://i.imgur.com/cWgip7b.png)
I also added Puzzle/Combat indicators for Unknown rooms, so you don't have to go through the layout to know its type (If not using the overlay)
![image](https://i.imgur.com/DBNP9Q7.png)

I also used Jasper-ketelaar's stream code since it is much cleaner, though not exactly as clean since I had to add another layer of ternary operators for the Room type indicators on Unknown rooms. I could have added another .map to keep it cleaner, but as I understand that would be quite inefficient.

I considered adding highlight colors for the different whitelists/blacklists that are implemented, however the clan chat only has a single highlight color, and I wouldn't be a fan of what the result would be. I could've added options to highlight either or many of the lists, but they would all have the same color... I'd also have to change all the code back to the for loop, in order to have more control on the chat colors/appending bit by bit. This is probably the final product for me, let me know what you guys think.