Follow GitHub templates for bug report and issue
Based on points made in #4160, update GitHub issue templates to be more
versatile, do not include any inline comments and be straightforward,
while pointing out reporter in correct direction. Also, one of the
crucial things that previous template for bug report was missing was
RuneLite and Launcher versions and steps to reproduce the bug.

Supersedes #4160

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>