Fix applying default config to session settings
cda98f moved the plugin manager to not be registered on the eventbus until
after the session was opened. The session config was relying on the plugin
manager applying the default config based on the session open event.

This moves session loading to be prior to applying the default config, so newly
added config options get properly set at startup.

Fixes config panel breaking which expects the config to always have non-null
values for config items with defaults.

Co-authored-by: Tomas Slusny <slusnucky@gmail.com>