Add overlay to Ring of Recoil
Initial commit to add a charges overlay to the Ring of Recoil.
Currently, there is no way to find out the source actor of hitsplats.
This implementation is based solely on assumptions. Namely, if the
player has been hit while wearing the ring of recoil, and the opponent
receives a hitsplat within 10 ticks which matches the calculated
reduction, charges are reduced from the ring. This works 90% fine on
Zulrah, but has issues on other events like Pest Control.

Implements #2109

Signed-off-by: Jaimy Smets <smets.jaimy@hotmail.com>