Implement Refocus plugin. Closes #5011
This plugin saves mouse position when the window focus is lost while the
mouse is over the viewport. It will restore this position when focus is
regained.

This is intended to relieve the strain caused by constantly moving the
mouse around while playing more than one RuneLite client simultaneously
(which is permissible under RuneScape rules).