Option to remove opponent info on bosses who already have a vanilla one
Added a toggle-able feature to disable health bars on bosses who have a vanilla one, the npc's that are blacklisted are in opponentinfooverlay.java all the way at the bottom.

Im not sure if there is a easier way to do this. If so tell me and ill close the pr.

Solves https://github.com/runelite/runelite/issues/5570