Opponentinfo removed if boss with healthbar
If a boss is defined to have a healthbar then it will not show up when attacked.
I looked into detecting if the healthbar widget is open but jagex seems to use multiple ones as well as doing this would not show other npc's health when in a boss room (TOB for example) thus making this solution around the same effort for a better result.

Fixes #5570