Option to show username when chat input is locked.
When a user enables the WASD camera controls, the chat input is locked and the user prompted to press 'Enter' to enter text into the chat input field. When not typing, the chatbox hides the username and as such it is sometimes confusing to know, at a glance, what user is logged in.
In this commit, a toggle has been added to prefix the chat prompt with the username.