Adding logic to ensure we don't add empty loot entry's
If a player uses a crate on a deposit box or dies with a crate, the
crate will be identified as used with 0 loot gained. We should ignore
these cases. There is a possible edge case where the player deposits
a supply crate and picks up an item in the WintertodtLoot enum and it
will still be identified as a loot entry.