XP Drops: Add support for many situations
The overlay is now created if needed when the client loads.
The overlay now properly goes away when the plugin is disabled.
The overlay now changes its default position properly when the client is changed to/from resizeable mode.
The overlay now changes its default position properly when the client's dimensions are changed in resizeable mode.
I'm on the fence about updateOverlay(), maybe it's a bit overkill for code that's reused only twice.