XP Drops: Add configs for moveable drops and recolouring drops
This caused way more issues than it had any right to, but it also highlighted some problems.

Adding onWidgetHiddenChanged makes it work with changing user accounts, as the widget is only ever loaded once.

Also, a lot of client functions cannot be called outside the main thread. onConfigChanged is not in the main thread. TIL

client.getViewportWidget().getBounds() can sometimes have a width/height of 0/0, meaning that while it's loaded, it's not set up properly. Adding a check for that makes it work.