Implement item information tooltip plugin!
How it works:
Currently this plugin is using osrsbox.com item database to retrieve all the item information via the REST API and applies them to the item tooltip. The item search is standoff so it won't show information for an item thats name isnt an exact match. It might be worth while transitioning this database over to the runelite http service db and relying on that. For now this works fine.

Demo: https://youtu.be/_LbPpwzwN1U