Add notification option for valuable and untradeable items
Also add fix to not send chat notification messages to the in game messages. This removes the duplicated game messages when a highlighted word is notified.
Simplify logic

Switch to checking for SERVER message instead of all messages.
Use overload instead of optional parameters


Contains to startsWith


Add more support for overloaded options.