Optimizations to playerindicator overlay
There is some overhead for fetching config values and it adds up fast in a render hook. Config values are now stored as class members of the overlays and are updated with a subscriber. Also moved some of the overlay enabled checks to the render function so it won't run the itterator if its not going to draw anything.