widget overlay: Add ctor to gate on child visibilities
This commit adds a constructor which allows widget overlays to be
gated upon visibility of other widgets. This is useful to help control
drawing of visible container widgets which may contain a number of child
widgets which may at some point all be hidden, thus drawing an overlay
for an empty bound.