Add option to require shift for overlay menus
Enabled by default. This was on original PR review and was discussed in
Discord. It is better to just use what we already use for adding our
custom entries at other places, and there is possibility that overlays
will have a lot of right click menus in future, so making it require
shift is most sane option.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>