Update slayer points in tooltip while in rewards
Previously, while in the slayer rewards menu, the points variable would
be updated on each tick to be equal to whatever was displayed in the
"Slayer points: " widget. However, the tooltip wasn't redrawn until the
end of the task, and therefore changes in points while in the rewards
menu would not be reflected. Now, if a change in slayer points is
detected while in the rewards menu, the slayer tooltip will be redrawn
and the change therefore reflected.