mlm plugin: fix duplicate ore counting
This commit aims to fix the duplication of ores as described in issue
7641. There were 2 bugs leading to duplication. First, the flag for
determining if the inventory should be checked for ores was being set
any time the ore count varbit was changed, regardless of if it increased
or decreased. Second, the plugin was checking the entire inventory
instead of computing a delta of the inventory changes causing ores to be
counted twice if you withdrew from the sack before banking.