Major update.
Fixed the issue where if you close the bank as you withdraw or deposit something it registers after the bank item container is closed.

Neatened up a lot of code.

Added in a check for if the option to withdraw noted is selected.

No *NEW* functionality has been added. It just works a lot better.