XpTracker Plugin Addition (Xp/Time Left swap)
Add an option to the XpTrackerConfig to allow the choice to swap the XP Left tracker with the Time Left tracker (this will also change the data in the progress bar tooltip as well).

The text has been formatted and shortened with some conditions to ensure that it shows the significant values.