Remove and edit wiki lookup conditions
Remove quick_prayer_group_id because i thought it was for the orb, i think it is for the quick prayer screen when clicking op the setup quick prayer button.

Change some text conditions to "Examine" for better placement of lookup wiki and consistancy.

Remove condition for SHOP_INVENTORY because it does not seem to work, items in the shop with SHOP_GROUP do work though. Also tried removing the widget conditioning and only look for "Examine" menu option, it works for all the items with an "Examine" option but not for the items in the shop inventory for some reason.