Update BA Plugins, Update from RL master (#226)
* xp tracker: display >1m exp with thousandths precision

* fishing plugin: fix trawler timer with one tick left

* Remove gray pixel from Dragon Scimitar cursor (#8725)

* client: add smelting plugin

Co-authored-by: Adam <Adam@sigterm.info>

* constants: add GAME_TICK_LENGTH and use where needed

Also update many usages of 20ms to CLIENT_TICK_LENGTH

* regenmeter: add option to notify before next hp regen

* barrows plugin: add prayer drain timer

* chat: ensure queued messages are always cleared after adding

Mark queued message type as non null because addChatMessage will throw
if it is. This was causing the client to get stuck trying to add the
same broken message each loop.

* party plugin: add partyinfo debug command

* party service: check joins are for the current party

The server can resume party membership on handshake to parties other
than the current client party (eg. if the client is restarted or for any
other reason the current party changes). This desyncs otherwise if the
joins aren't checked to be in the current party.

* Updated barbassault/ba tools

* Update