combat level plugin: Fix next combat calculations
The existing calculations were both incorrect for range and magic level
calculation, which is demonstrated in the added test case, and were
complicated to the point that few developers were able to understand how
the plugin achieved its results.

This patch changes the plugin to use the same straightforward approach
which is used by other known correct combat calculators such as the
tip.it calculator, which determines the number of levels to advance a
combat level by calculating the combat level while incrementing the
specified skill until it would result in a higher combat level.

Fixes runelite/runelite#7411