itemskeptondeath: fix many edge cases and add tests
This adds support for most degradeable and imbued items, which have
death prices that are fixed offsets of the prices of some other
reference item, clue boxes, and jewelery.

The plugin was refactored largely to be able to add the tests, and now
works on our internal DeathItem and ItemStack classes instead of widgets
and items.

Co-authored-by: Adam <Adam@sigterm.info>