Stops the overlay from activating when it's not supposed to.
Before this, you could for example get this bug:
Have player safespots turned on and NPC safespots turned off.
Interact with another player, which creates a safespotlist.
After this, whenever you're interacting with an NPC and then stop
interacting with it, the overlay would turn on with the tiles from the
old playerinteraction safespotlist highlighted for 10 seconds.