Remove runescape-api and mixins
This code is used by internal tooling to build the client patch, which
is already closed source, and is not directly shipped as a dependency.

Outside contributions to it are rare due to the lack of build tools, and it is
primarily being abused by others. As such, I no longer wish to release
further changes to it under BSD-2.