Status Orb Plugin: Fix run energy overlay
Owain broke the overlay on the run energy orb by adding an extra if to the logic where it wasn't really needed so that it only ended up showing the overlay when the `run time left` option was enabled. This fixes it back to how it was originally where it always shows the overlay whenever you mouse over the run energy orb.