Revert "Shutdown plugins on window close"
This reverts commit 6e080949f9943ac118b62840bac0624f1f3e6d7b.

This never worked because it ran stopPlugin on the AWT thread
which calls invokeAndWait, which throws when not on the awt
thread. Also stopPlugin only attempts to stop plugins that are
marked as disabled, but it is only called for plugins which are
enabled, so nothing actually happens. This also creates a deadlock
if you try to close the client while starting it, because
startPlugin is locked on one thread, but stop plugin is trying
to take the lock on the edt.