Fix wrong rock respawn timers
Fix respawn timers for
 - Adamant rocks in Wilderness Resource area
 - Coal rocks in Miscellania
The doubled respawn rate of the adamant rocks is a possible Jagex bug as
they are the only type of rock within the Wilderness resource area to do so.
The coal rocks in Miscellania are mined for favour instead of ores
and therefore have a shortened respawn time.