opponentinfo: Track NPCs which spawn and attack simultaneously
When being attacked by an NPC which spawned on the same tick it attacks
(an example being the Double agent when doing emote clues), if a player's
auto-retaliate is on, the plugin would not properly identify the
NPC as being attacked because it would be `null` when the player's
interacting changed. This is fixed by reassigning the last tracked
opponent to the current interacted NPC if it was null.

Fixes runelite/runelite#6425