Add Abyssal Sire Plugin
Add Abyssal Sire Plugin for toggling left-click to walk under Spawns and Scions.

This plugin improves the QoL of fighting the Abyssal Sire by de-prioritizing the Attack option on the Spawns and Scions, allowing you to left-click walk under them. This plugin solves a similar problem for Abyssal Sire that the "left click walk on core" plugin option does for Corporeal Beast.