object markers: use id instead of name to identify objects
With the addition of the ability to mark WallObjects some normal object
markers created unremovable artifacts when hopping. This is basically
because when getting the ObjectComposition of many objects they have the name
"null" even some walls. Namely, when marking the top part of the ladder to
upper level mlm, which has the name "null". When we then log in again the
wall just below it is a part of the same tile with the same name "null"
which adds it to the highlighted objects of the region. Since we can never
select that wall again, we can never unmark it.