xp tracker: enable xp progress tooltips for p2p skills on f2p worlds
The information is already available if you look it up on the hiscores
and the members only skills already have changed opacity to indicate that
they are members only. Only tooltips for skills with >0 xp will be overridden
so that free players who have never been members will get the intended F2P feeling.