Clarified a few comments and minor optimization
Some comments were outdated due to moving around code, fixed that. Slightly optimized the removal of the panel navButton in onGameTick depending on the location since we don't always need to check it. Also a potential bug fix in the onStart since I was checking the player location when potentially logged out, although I've never seen the bug occur.