Adding Zeah Runecraft Plugin
This plugin will highlight the dense runestones with certain colours to provide context as to what actions can be taken.

When a dense runestone is depleted, it is highlighted red.
When the player is currently mining a dense runestone, the runestones are highlighted orange.
When the player is not currently mining, the runestones are highlighted green.
When a depleted runestone becomes available again, it will either be highlighted orange or green, depending on if the player is mining or not.

The highlighting can be diabled by the new config group "Zeah Runecraft".
I debated whether to combine it with the existing "Runecraft" config group, but this looks very Abyss specific.

The purpose of this plugin is to provide a better experience when AFKing as it can be easy to miss sound cues when a rock depletes.