inventorygrid: use mouse dragged distance to determine if overlay should show
Replace the time dependant showing of the inventory overlay with a distance
based one. An item sprite only follows the mouse if the mouse is held
down and dragged ~5 pixels. Thus, the overlay will only show when you are
certain to be dragging an item and the time-delay is unnecessary.

Original PR by Jbleezy.