Allowing anti-drag key to be configurable
Currently the plugin only supports the shift key for increasing the drag delay but this isn't ideal as shift is often bound to "shift to drop", meaning that items can be accidently dropped if this option is enabled.

I've added a new ModiferlessKeybind config option to allow the user to choose any key that they wish to be used in place of shift.