inventorygrid: re-add a delay to showing the overlay
If a player for example tries to drop an entire inventory quickly with the
overlay turned on, the inventory grid will flash for each small mouse movement
if the left mouse key is held down. This change mitigates the issue by adding a
delay option that controls the amount of time the user has to wait after
pressing it before enabling any overlay functionality. Showing the grid still
requires the mouse to be moved ~5px away from the first clicked position.

This delay just delays the activation of the overlay, as the the delay
countdown is started upon item press rather than the exit this ~5px range.
Making it respect the exit is needlessly complex as desired functionality is
reached with the current changeset.