clue plugin: add light requirements to clues
This commit adds a warning message to the clue scroll overlay when a
player has a clue requiring a light source and does not have a light
source equipped or in their inventory. It adds logic to check for the
status of fire pits in areas which would remain permanently lit that
way.

Closes runelite/runelite#9626

Co-authored-by: David <Davidlomath@hotmail.co.uk>
Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>
Co-authored-by: Adam <Adam@sigterm.info>