kc chat command: Use hiscore values for supported activities
If the activity is not supported or the hiscore request times out
the RL config value will be used. If the request succeeds it will show
the rank and highest level between the hiscore and RL values.
If the RL value is lower it will be updated to match the hiscores.