runenergy: update graceful recovery rate logic
Updates the energy recovery rate calculation to reflect the changes made
in the 9th January 2020 game update.

Each individual piece of graceful now boosts your recovery rate by 3%,
with an extra 1% for the top and legs. With the full set, an extra boost
of 10% is added, totalling 30%. Previously, there was no boost for
individual pieces, instead the full set was required for a flat boost of
30%.