Remove bounty hunter emblem tier from player name for lookup
A player in the wilderness of a bounty hunter world with a bounty hunter
emblem in their inventory will have their name decorated in the player
menu with an emblem icon and the tier of emblem they are carrying. Before
removing any tags, that name will look like this:

<col=ffffff>bfmoatbio<img=20>10<col=00ff> (level-126)

where "<img=20>10" represents the emblem icon (<img=20>) and tier (10).
removeTags will remove the icon tag but leave the emblem tier, resulting
in an incorrect username returned by getMenuTarget (in the case of the
example, bfmoatbio10).

Before calling removeTags on a player menu target, perform a special
remove of the bounty hunter emblem icon and its tier, if present.

Change the MenuManager constructor to package-private (was private) for
use in new tests, and mark with @VisibleForTesting for documentation.