banktags: remove duplication of widgets when staying within bank
From one of my previous commits, in the scenario where you start in the
main bank interface, click on either gear or settings, then return to
main bank interface, the code created duplicate widgets for upButton,
downButton, newTab and all tag tabs. This error was not visible unless
you right clicked on said widgets and found duplicate settings. The
change introduces an if block that differentiates the case of opening
the bank, and the case of switching among bank interfaces and returning
to the main bank interface. This should get rid of duplicates without
side effects.