banktags: reassign trigger to recall tag tab when staying within bank
As per the suggestion, I changed the trigger that recalls the last tag
tab when switching interfaces within the bank, from checking that the
main bank interface is not hidden, to checking that one of two buttons
has been clicked, "Hide worn items" and "Hide menu". This must use the
invokeLater method because clicking either button changes the state of
isHidden but the client must complete that process before recalling the
saved tag tab. The behavior should remain the same as the last commit.