xpdrops: fix recoloring xpdrops when prayer is flicked on the previous tick
Previously the widget hidden event would not run until the next client
tick, due to it being on a timer, and the game tick event being deferred
until the next frame. Now that we are hooking to the script being
executed it fires prior to the game tick event being fired.