bank tags: invoke only the scrollbar update after removing separators
Invoking the tab separator removal allowed tab separators to be visible and interactable for a frame. On slower computers, users could click and navigate to a tab during that period. The removal no longer needs to be invoked after e4b43d9 so we can just invoke the script run.