Optimize SceneUploader.upload(Scene scene)
Remove the reset() iteration across the tiles when uploading a scene, the upload() iteration itself is enough to set the buffer offsets of every object. The sceneId is now used to check if a model has already been uploaded or not. A test of loading the center of Varrock 1000 times shows a small improvement from 33 to 28 seconds.