fix incorrect drop rate coloring
The coloring for the loot potential on the main barrows table is incorrect.  I assume it was meant to turn yellow when death runes were added to the loot potential, green when blood runes were added, and white again for bolt racks.  But the numbers are off by just a bit.  See wiki percentages:
https://oldschool.runescape.wiki/w/Chest_(Barrows)#Main_table