skill calc: add 4 dose potions
Add 4 dose Extended Antifires and Anti-venoms to the herblore skill calculator. These potions are able and often are made using 4 doses at a time.

Super Antifires and Anti-venom+'s can not be made using 3 doses at a time, so were fixed to display the right potion dose (4). The xp for these potions was already correct.