Update CannonSpots.java
Removed coordinates for the elf/mourner cannon spot located under the mourner hq. After the release of the quest Song of The Elves you are still able to place a cannon down here but the cannon will not fire regardless if you have completed the quest or not therefore the cannon marker is not valid.