gpu: workaround for forced anti-aliasing
Forced anti-aliasing by the GPU applies multisampling to the default
framebuffer, and when blitting between two multisampled FBOs, the number
of samples in both FBOs has to match.

In order to make sure our anti-aliasing FBO has the same number of
samples as the default FBO, we bind the default FBO and check whether
its number of samples is non-zero. If it is, use this as the number of
samples for our anti-aliasing FBO instead of getting it from the
configured anti-aliasing mode.

Co-authored-by: Adam <Adam@sigterm.info>