I've moved relevant variables that can be local into the null check block. set the default price label to null.
I've added and injected ClientThread var in the Plugin, added a ClientThread var to the constructors of the SkillCalculatorPanel and SkillCalculator classes and passed the ClientThread (SkillCalculatorPlugin -> SkillCalculatorPanel -> SkillCalculator) via them to be able to use it in the SkillCalculator calculate function.

I invoke the price calculating code block with the client thread inside after passing the nullcheck  (so it only invokes when it needs to actually access the getItemPrice function).
added an "effectively" final ActionCount variable to be used inside the invoked block.