banktags: Fix bank scroll height with separators hidden
Prior to this change, an extra row of empty space was added to tag tabs
with items reaching the end of a row (ie. multiples of 8) which led to
scroll behavior not matching vanilla for tabs with enough items to
require scrolling. This change fixes this behavior to add extra scroll
height only when a new row is started.