banktags: Add missing bottom height with separators hidden
In addition to the 36 pixels of height allotted to each bank item row,
an additional 4 pixels of height exists at the bottom of the bank item
container. This change ensures that a bank tag scrolled to the bottom
will have its items displaying at the exact same y values as that of a
normal bank tab.