runelite-client: use menu entries for focusing world map points
The previous way had usability issues since it would cover up the close
button, consumed clicks for no reason on any icon, making it difficult
to pan around the map when zoomed out, and had no tooltips.

It previously also accessed widgets on the EDT, which is a threading
issue.