runelite: Fix login screen mute button bounds
The login screen mute button is only strictly bounded on two sides. This
was fine before resizable but now the size of the button is silly.
Add a workaround to consume clicks that are in this area and outside the
fixed client dimensions. AS there is no way to determine whether the
client is in the world hopper screen and the next button anchorings
to the side of the client canvas a small area must be allowed to remain
to avoid nulling clicks on this button.