special counter: ignore vet'ion hellhounds
Prevents spec counter resetting when attacking a skeleton hellhound spawned by Vet'ion. The hellhounds are spawned at 50% HP, and are required to kill to continue the fight, so the player will continue fighting Vet'ion after they are killed.

It's possible that a player may want to use a spec against the hellhounds, but I'm not aware of any reason for it.