chat message manager: don't apply color changes to message nodes
Apply the color changes at chat build time directly to the message being
set on the component, instead of prepending the color change onto the
message itself.

This is a bit more robust since it doesn't break things which depend on
the chat message not starting with a color, which they would not
otherwise, such as the chat filter.