Add @SafeVarargs to ExternalPluginManager.loadBuiltin
Fixes warning "Unchecked generics array creation for varargs parameter" on
loadBuiltin. This warning is trying to prevent a RuntimeException which
happens iff the varargs are modified in a particular way. Varargs are not
modified, which is guaranteed by PluginManager.loadPlugins using a Stream.

Change loadPlugins to take a Stream of Plugins. Saves a few allocations.

Allow loading of plugins that are not direct subclasses of Plugin, but
subclasses of subclasses of Plugin, etc. (fix in loadPlugins) and add a
test case that this works

Fix an iterator invalidation bug in PluginManager.topologicalSort.
Graph.sucessors is an iterator; removing nodes while iterating throws
a concurrentModificationException. Fix: add an allocation.

Fill out existing javadocs for topologicalSort