Add commands for !quests !collections !cbtasks Change submitDiaries and getDiaries to submitCharacterSummary and getCharacterSummary to be more generic to use with other commands.
Add csTotal (character summary total) method to find total number of Acheivements, Quests, Combat Tasks and Collections to make the message look nicer using the widgets.

To do: Still hijacking the qp method from ChatClient/ChatCommandsPlugin for sending and receiving ints with the http-service instead of using their own. Not sure if this is necessary but would look cleaner and less confusing.