timetracking: compost tracking
Implements compost tracking for farming patches using a similar concept as the Examine plugin.

The plugin watches for composting and inspect actions, then stores a pending compost which is finalized by chat messages. In order to distinguish between multiple compost actions queued, the tracker also checks for adjacency to the patch and has a timeout of 30 seconds.