hunter: add dead box trap detection
Box traps work by having the trap interact with a chin, which causes the chin
to interact with the trap. If a trap interacts with a chin, but the chin
interacts with a different trap and gets caught, the original trap is
stuck in a dead state where it cannot catch any further chins.

This detects the dead state and changes the overlay color to reflect
that.

Co-authored-by: Tjieco <Tjiecodonselaar@hotmail.nl>