Add an option to draw a pie progress indicator on fishing spots.
This is a QoL change to make fishing even AFKer; players will be able to clearly see which nearby fishing spots are the youngest.
This change extends the existing functionality for minnow fishing spots, in which a pie progress indicator is drawn to show its age, to other fishing spot types.
This change preserves the existing behavior for minnows and keeps the minnow-specific options while adding a new option to enable spot timers for other spots.

Additionally, this change adds an overlay "warning" color option that the timer fades to as the spot ages.
This uses the existing "20% of the remaining time left" as the point for the full warning color, since that's what the minnow options were doing with their hard-coded warning color.

Additionally this change makes the pie indicator smaller, draws it to the bottom-right of the spot icon, and makes it the last overlay to be rendered so that it isn't covered by the spot icon.
Without this change it is difficult to see how much progress is on the timer; making it icon smaller and drawn in the bottom-right also minimizes occlusion of the spot icon.

This change reuses the existing logic with MinnowSpot, but renames it to TrackedFishingSpot to better indicate that it no longer tracks only minnow spots.
This change also adds new debug logging to make it easier to measure fishing spot timers in the future.
For example, this logging was used to show that, despite the claims on the wiki, anglerfish spots appear to last for the same duration as other fishing spots.
Karambwanji should probably be measured at some point since those spots seem to last a very long time (not karambwan spots, which last forever).