npc indicators: Track respawns only for dead NPCs
Some NPCs such as Giant Mole have mechanics where they can spawn in a
predictable location, but also move in a large distance (possibly beyond
that which the server transmits to the player) to cause despawn events
during their fight. Previously, this would cause a respawn indicator to
appear at their spawn location each time such a despawn occurred. This
commit changes that behavior to only affect NPCs which are dead at the
time of despawning.