camera: rotate camera even if there are menu entries
Changes the rightClickMovesCamera and middleClickMenu configs to always send an extra middle click
event so that the camera can be rotated regardless if there are menu entries or not.

This has to be done in a bit of a convoluted way as the client only rotates the camera if the order of
operations is  (1) right click, (2) middle click. As the middle click event has to be dispatched after
the right click event, it is done in the drag event handler.