barrows: chat message when looting chest updated
The chat message is now of type GAMEMESSAGE, the items are included in the chat message, and the items are sorted from greatest to least G.E. value. The format of the items in the chat messages follows that of the items in the chat messages of valuable drops for loot drop notifications when killing monsters: "{item} ({value})" if quantity = 1 and "{quantity} x {item} ({value})" if quantity > 1.